Fix issues with search see feature/search.feature Indentation

+++ admin/photos_add_direct.php
@@ -22,40 +22,34 @@
 // | USA.                                                                  |
 // +-----------------------------------------------------------------------+
 
-if (!defined('PHOTOS_ADD_BASE_URL'))
-{
-  die ("Hacking attempt!");
+if (!defined('PHOTOS_ADD_BASE_URL')) {
+    die ("Hacking attempt!");
 }
 
 // +-----------------------------------------------------------------------+
 // |                        batch management request                       |
 // +-----------------------------------------------------------------------+
 
-if (isset($_GET['batch']))
-{
-  check_input_parameter('batch', $_GET, false, '/^\d+(,\d+)*$/');
+if (isset($_GET['batch'])) {
+    check_input_parameter('batch', $_GET, false, '/^\d+(,\d+)*$/');
 
-  $query = '
-DELETE FROM '.CADDIE_TABLE.'
-  WHERE user_id = '.$user['id'].'
-;';
-  pwg_query($query);
+    $query = 'DELETE FROM '.CADDIE_TABLE.' WHERE user_id = '.$user['id'].';';
+    pwg_query($query);
 
-  $inserts = array();
-  foreach (explode(',', $_GET['batch']) as $image_id)
-  {
-    $inserts[] = array(
-      'user_id' => $user['id'],
-      'element_id' => $image_id,
-      );
-  }
-  mass_inserts(
-    CADDIE_TABLE,
-    array_keys($inserts[0]),
-    $inserts
+    $inserts = array();
+    foreach (explode(',', $_GET['batch']) as $image_id) {
+        $inserts[] = array(
+            'user_id' => $user['id'],
+            'element_id' => $image_id,
+        );
+    }
+    mass_inserts(
+        CADDIE_TABLE,
+        array_keys($inserts[0]),
+        $inserts
     );
 
-  redirect(get_root_url().'admin.php?page=batch_manager&filter=prefilter-caddie');
+    redirect(get_root_url().'admin.php?page=batch_manager&filter=prefilter-caddie');
 }
 
 // +-----------------------------------------------------------------------+
